feat!: implement protected endpoints for users and authorization logic

+++ app/users/service.py
@@ -4,9 +4,10 @@ from uuid import UUID
 from fastapi import Depends
 from pydantic import EmailStr
 
+from app.auth.exceptions import PermissionDeniedException
 from app.users.exceptions import EmailAlreadyRegisteredException, UserNotFoundException
 from app.users.repository import UserRepository
-from app.users.schemas import UserCreate, UserResponse, UserUpdate
+from app.users.schemas import UserCreate, UserPrivate, UserPublic, UserUpdate
 
 if TYPE_CHECKING:
     from app.users.models import User
@@ -16,50 +17,62 @@ class UserService:
     def __init__(self, user_repository: Annotated[UserRepository, Depends()]) -> None:
         self._user_repository: UserRepository = user_repository
 
-    async def list_users(self) -> list[UserResponse]:
+    async def list_users(self) -> list[UserPublic]:
         users = await self._user_repository.find_all()
-        return [UserResponse.model_validate(user) for user in users]
+        return [UserPublic.model_validate(user) for user in users]
 
-    async def get_user_by_id(self, user_id: UUID) -> UserResponse:
+    async def get_user_by_id(self, user_id: UUID) -> UserPublic:
         user = await self._user_repository.find_by_id(user_id)
         if not user:
             raise UserNotFoundException(str(user_id))
 
-        return UserResponse.model_validate(user)
+        return UserPublic.model_validate(user)
 
     async def get_user_by_email_internal(self, email: EmailStr) -> User | None:
         return await self._user_repository.find_by_email(email)
 
-    async def get_user_by_email(self, email: EmailStr) -> UserResponse:
+    async def get_user_by_email(self, email: EmailStr) -> UserPublic:
         user = await self.get_user_by_email_internal(email)
         if not user:
             raise UserNotFoundException(value=email, attribute="email")
 
-        return UserResponse.model_validate(user)
+        return UserPublic.model_validate(user)
 
     async def is_email_already_registered(self, email: EmailStr) -> bool:
         user = await self._user_repository.find_by_email(email)
         return user is not None
 
-    async def create_user(self, create_data: UserCreate) -> UserResponse:
+    async def create_user(self, create_data: UserCreate) -> UserPublic:
         if await self.is_email_already_registered(create_data.email):
             raise EmailAlreadyRegisteredException(create_data.email)
 
         user = await self._user_repository.create(create_data)
-        return UserResponse.model_validate(user)
+        return UserPublic.model_validate(user)
 
-    async def update_user(self, user_id: UUID, update_data: UserUpdate) -> UserResponse:
+    async def update_user(
+        self, user_id: UUID, update_data: UserUpdate, actor_id: UUID
+    ) -> UserPrivate:
         user = await self._user_repository.find_by_id(user_id)
         if not user:
             raise UserNotFoundException(str(user_id))
+
+        if user_id != actor_id:
+            raise PermissionDeniedException(
+                action_description="Modify someone else's profile."
+            )
 
         await self._user_repository.update(user, update_data)
-        return UserResponse.model_validate(user)
+        return UserPrivate.model_validate(user)
 
-    async def delete_user(self, user_id: UUID) -> None:
+    async def delete_user(self, user_id: UUID, actor_id: UUID) -> None:
         user = await self._user_repository.find_by_id(user_id)
         if not user:
             raise UserNotFoundException(str(user_id))
+
+        if user_id != actor_id:
+            raise PermissionDeniedException(
+                action_description="Delete someone else's profile."
+            )
 
         await self._user_repository.delete(user)
